adjust robot abstraction
